Replace hard-deletes by soft-deletes to maintain document history
Today we can use the soft-deletes feature from Lucene to maintain a
history of a document. This change simply replaces hard-deletes by
soft-deletes in Engine.

Besides marking a document as deleted, we also index a tombstone
associated with that delete operation. Storing delete tombstones allows
us to have a history of sequence-based operations which can serve in
recovery or rollback.

+++ server/src/main/java/org/elasticsearch/index/engine/InternalEngine.java
@@ -1220,7 +1220,14 @@ public class InternalEngine extends Engine {
             if (plan.currentlyDeleted == false) {
                 // any exception that comes from this is a either an ACE or a fatal exception there
                 // can't be any document failures  coming from this
-                indexWriter.deleteDocuments(delete.uid());
+                if (softDeleteEnabled) {
+                    final ParseContext.Document tombstone = engineConfig.getMetaDocSupplier().newMetaDoc(
+                        delete.type(), delete.id(), plan.seqNoOfDeletion, delete.primaryTerm(), plan.versionOfDeletion);
+                    tombstone.add(softDeleteField);
+                    indexWriter.softUpdateDocument(delete.uid(), tombstone, softDeleteField);
+                } else {
+                    indexWriter.deleteDocuments(delete.uid());
+                }
                 numDocDeletes.inc();
             }
             versionMap.putUnderLock(delete.uid().bytes(),
